+++ app/src/main/java/com/tosslab/jandi/app/ui/message/v2/MessageListV2Presenter.java
@@ -1081,7 +1081,7 @@ public class MessageListV2Presenter {
 
     }
 
-    public void updateStarredOfMessage(int messageId, boolean starred) {
+    public void updateStarredOfMessage(long messageId, boolean starred) {
         int index = adapterModel.indexByMessageId(messageId);
         adapterModel.modifyStarredStateByPosition(index, starred);
         view.refreshMessages();
